Faces v+1- S'occuper des problèmes de collages. (copier-coller) -> Resolved. 2- Améliorer "Face" (...) v-3) try to repair ChooseEffectsActivity2 Ideal: une liste de liste de FileDetail (int order, currentFile (may be null  to work in memory, current, pixM, currentPixMZoom, intentExtra?) -- peut-être voir items/details fragments. Auto reduce image size and 3 bytes color instead of 3 double color for calculation.

+++ app/src/main/java/one/empty3/feature20220726/ProcessEval.java
@@ -33,7 +33,7 @@ public abstract class ProcessEval {
     private ArrayList<File> files = new ArrayList<>();
     private ArrayList<PixM> pixMaps = new ArrayList<>();
     private ArrayList<double[]> features = new ArrayList<>();
-    private double[] x;
+    private byte[] x;
     private double[] y;
     private double[] w;
     private double[] w1;
@@ -86,7 +86,7 @@ public abstract class ProcessEval {
         return a;
     }
 
-    private double[] dotOuter(double[] x, double[] w) {
+    private double[] dotOuter(byte[] x, double[] w) {
         double[] a = new double[x.length];
         for (int i = 0; i < a.length; i++)
             a[i] = x[i] * w[i];
